feat: lock
