+++ Sources/Notifly/notifly-ios-sdk/notifly-ios-sdk/SourceCodes/Infrastrucutres/InAppMessage/InAppMessage+Constant.swift
@@ -23,4 +23,5 @@ enum InAppMessageConstant {
     }));
         });
     """
+    static let inAppMessageChannel = "in-app-message"
 }
