+++ Sources/Notifly/notifly-ios-sdk/notifly-ios-sdk/SourceCodes/Infrastrucutres/Tracking/TrackingManager.swift
@@ -94,25 +94,26 @@ class TrackingManager {
                isInternal: Bool,
                segmentationEventParamKeys: [String]?)
     {
-        var trackingEventName: String
-        if isInternal {
-            trackingEventName = "notifly__" + eventName
-        } else {
-            trackingEventName = eventName
+        guard let notifly = try? Notifly.main else {
+            Logger.error("Fail to track Event. \(eventName)")
+            return
         }
+        let trackingEventName = NotiflyHelper.getEventName(event: eventName, isInternalEvent: isInternal)
+        let userID = (try? notifly.userManager.getNotiflyUserID()) ?? ""
+
+        let externalUserID = notifly.userManager.externalUserID
         let currentTimestamp = AppHelper.getCurrentTimestamp()
-        guard let syncStateFinishedPub = try? Notifly.main.inAppMessageManager.syncStateFinishedPub else {
+        guard let waitPub = try? Notifly.waitPub else {
             Logger.error("Fail to track Event. \(trackingEventName)")
             return
         }
-
-        syncStateFinishedPub.flatMap { _ in
+        waitPub.flatMap { _ in
             try? Notifly.main.inAppMessageManager.updateEventData(eventName: trackingEventName, eventParams: eventParams, segmentationEventParamKeys: segmentationEventParamKeys)
             return self.createTrackingRecord(eventName: eventName,
                                              eventParams: eventParams,
                                              isInternal: isInternal,
                                              segmentationEventParamKeys: segmentationEventParamKeys,
-                                             currentTimestamp: currentTimestamp)
+                                             currentTimestamp: currentTimestamp, userID: userID, externalUserID: externalUserID)
         }
         .sink(receiveCompletion: { completion in
                   if case let .failure(error) = completion {
@@ -133,7 +134,7 @@ class TrackingManager {
                               eventParams: [String: Any]?,
                               isInternal: Bool,
                               segmentationEventParamKeys: [String]?,
-                              currentTimestamp: Int) -> AnyPublisher<TrackingRecord, Error>
+                              currentTimestamp: Int, userID: String, externalUserID: String?) -> AnyPublisher<TrackingRecord, Error>
     {
         guard let notifly = try? Notifly.main else {
             return Fail(outputType: TrackingRecord.self, failure: NotiflyError.notInitialized)
@@ -144,16 +145,21 @@ class TrackingManager {
                 .eraseToAnyPublisher()
         }
 
+        guard let notiflyDeviceID = AppHelper.getNotiflyDeviceID(),
+              let deviceID = AppHelper.getDeviceID(),
+              let appVersion = AppHelper.getAppVersion(),
+              let sdkVersion = NotiflyHelper.getSDKVersion()
+        else {
+            Logger.error("Failed to track event: " + eventName)
+            return Fail(outputType: TrackingRecord.self, failure: NotiflyError.unexpectedNil("Device data is invalid."))
+                .eraseToAnyPublisher()
+        }
+
         return deviceTokenPub.tryMap { pushToken in
-            let userID = (try? notifly.userManager.getNotiflyUserID()) ?? ""
-            if let notiflyDeviceID = AppHelper.getNotiflyDeviceID(),
-               let deviceID = AppHelper.getDeviceID(),
-               let appVersion = AppHelper.getAppVersion(),
-               let sdkVersion = NotiflyHelper.getSDKVersion(),
-               let data = TrackingData(id: UUID().uuidString,
+            if let data = TrackingData(id: UUID().uuidString,
                                        name: eventName,
                                        notifly_user_id: userID,
-                                       external_user_id: notifly.userManager.externalUserID,
+                                       external_user_id: externalUserID,
                                        time: currentTimestamp,
                                        notifly_device_id: notiflyDeviceID,
                                        external_device_id: deviceID,
@@ -167,7 +173,7 @@ class TrackingManager {
                                        sdk_version: sdkVersion,
                                        sdk_type: NotiflyHelper.getSDKType(),
                                        event_params: AppHelper.makeJsonCodable(eventParams)) as? TrackingData,
-               let stringfiedData = try? String(data: JSONEncoder().encode(data), encoding: .utf8)
+                let stringfiedData = try? String(data: JSONEncoder().encode(data), encoding: .utf8)
 
             {
                 return TrackingRecord(partitionKey: userID, data: stringfiedData)
@@ -182,8 +188,6 @@ class TrackingManager {
         }
         .eraseToAnyPublisher()
     }
-
-    // MARK: - Private Methods
 
     private func setup() {
         // Submit the tracking event to API & log result.
