+++ Sources/Notifly/notifly-ios-sdk/notifly-ios-sdk/SourceCodes/Notifly/Notifly+Helper.swift
@@ -9,6 +9,10 @@ import Foundation
 
 @available(iOSApplicationExtension, unavailable)
 enum NotiflyHelper {
+    static func getEventName(event: String, isInternalEvent: Bool) -> String {
+        return isInternalEvent ? "notifly__" + event : event
+    }
+
     static func getSDKVersion() -> String? {
         return Notifly.sdkVersion
     }
